% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cutFunctions.R
\name{cut2byMedian}
\alias{cut2byMedian}
\title{recode a continous vector into a factor vector of two categories by median}
\usage{
cut2byMedian(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
a factor vector of two levels: low and high
}
\description{
recode a continous vector into a factor vector of two categories by median
}

