% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cutFunctions.R
\name{cut3byQuantile}
\alias{cut3byQuantile}
\title{recode a continous vector into a factor vector of three categories by quantile}
\usage{
cut3byQuantile(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
a factor vector of three levels: low, mid, high
}
\description{
recode a continous vector into a factor vector of three categories by quantile
}

