% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cutFunctions.R
\name{cut3bySD}
\alias{cut3bySD}
\title{recode a continous vector into a factor vector of three categories by sd}
\usage{
cut3bySD(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
a factor vector of three levels: low, mid, high
}
\description{
recode a continous vector into a factor vector of three categories by sd
}

