% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cutFunctions.R
\name{cut3equal}
\alias{cut3equal}
\title{recode a continous vector into a factor vector of three categories by low 1/3, mid 1/3, high 1/3}
\usage{
cut3equal(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
a factor vector of three levels: low, mid, high
}
\description{
recode a continous vector into a factor vector of three categories by low 1/3, mid 1/3, high 1/3
}

